+++ packages/vm/src/testing/mock.rs
@@ -10,7 +10,9 @@ use super::storage::MockStorage;
 use crate::backend::unwrap_or_return_with_gas;
 use crate::{Backend, BackendApi, BackendError, BackendResult, GasInfo};
 
-pub const MOCK_CONTRACT_ADDR: &str = "cosmwasmcontract"; // TODO: use correct address
+pub const MOCK_CONTRACT_ADDR: &str =
+    "cosmwasm1jpev2csrppg792t22rn8z8uew8h3sjcpglcd0qv9g8gj8ky922tscp8avs";
+
 /// Default gas multiplier in wasmd.
 /// See https://github.com/CosmWasm/wasmd/blob/v0.51.0/x/wasm/types/gas_register.go#L34
 const WASMD_GAS_MULTIPLIER: u64 = 140_000;
@@ -216,12 +218,62 @@ fn validate_length(bytes: &[u8]) -> Result<(), BackendError> {
     }
 }
 
-/// Returns a default environment with height, time, chain_id, and contract address
+/// Returns a default environment with height, time, chain_id, and contract address.
 /// You can submit as is to most contracts, or modify height/time if you want to
 /// test for expiration.
 ///
 /// This is intended for use in test code only.
+///
+/// The contract address uses the same bech32 prefix as [`MockApi`](crate::testing::MockApi). While
+/// this is good for the majority of users, you might need to create your `Env`s
+/// differently if you need a valid address using a different prefix.
+///
+/// ## Examples
+///
+/// Create an env:
+///
+/// ```
+/// # use cosmwasm_std::{Addr, BlockInfo, ContractInfo, Env, Timestamp, TransactionInfo};
+/// use cosmwasm_vm::testing::mock_env;
+///
+/// let env = mock_env();
+/// assert_eq!(env, Env {
+///     block: BlockInfo {
+///         height: 12_345,
+///         time: Timestamp::from_nanos(1_571_797_419_879_305_533),
+///         chain_id: "cosmos-testnet-14002".to_string(),
+///     },
+///     transaction: Some(TransactionInfo { index: 3 }),
+///     contract: ContractInfo {
+///         address: Addr::unchecked("cosmwasm1jpev2csrppg792t22rn8z8uew8h3sjcpglcd0qv9g8gj8ky922tscp8avs"),
+///     },
+/// });
+/// ```
+///
+/// Mutate and reuse environment:
+///
+/// ```
+/// # use cosmwasm_std::{Addr, BlockInfo, ContractInfo, Env, Timestamp, TransactionInfo};
+/// use cosmwasm_vm::testing::mock_env;
+///
+/// let env1 = mock_env();
+///
+/// // First test with `env1`
+///
+/// let mut env2 = env1.clone();
+/// env2.block.height += 1;
+/// env2.block.time = env1.block.time.plus_seconds(6);
+///
+/// // `env2` is one block and 6 seconds later
+///
+/// let mut env3 = env2.clone();
+/// env3.block.height += 1;
+/// env3.block.time = env2.block.time.plus_nanos(5_500_000_000);
+///
+/// // `env3` is one block and 5.5 seconds later
+/// ```
 pub fn mock_env() -> Env {
+    let contract_addr = MockApi::default().addr_make("cosmos2contract");
     Env {
         block: BlockInfo {
             height: 12_345,
@@ -230,7 +282,7 @@ pub fn mock_env() -> Env {
         },
         transaction: Some(TransactionInfo { index: 3 }),
         contract: ContractInfo {
-            address: Addr::unchecked(MOCK_CONTRACT_ADDR),
+            address: Addr::unchecked(contract_addr),
         },
     }
 }
@@ -248,6 +300,12 @@ pub fn mock_info(sender: &str, funds: &[Coin]) -> MessageInfo {
 mod tests {
     use super::*;
     use cosmwasm_std::coins;
+
+    #[test]
+    fn mock_env_matches_mock_contract_addr() {
+        let contract_address = mock_env().contract.address;
+        assert_eq!(contract_address, Addr::unchecked(MOCK_CONTRACT_ADDR));
+    }
 
     #[test]
     fn mock_info_works() {